restructured validation, added kubconfig to .gitignore
